Refactor: Split AgentServer
