+++ server/src/GameServer/Connection/AgentServer.MessageParser.cs
@@ -1,0 +1,127 @@
+using System.Text.Json;
+
+namespace GameServer.Connection;
+
+public partial class AgentServer
+{
+
+    /// <summary>
+    /// Parse the message
+    /// </summary>
+    /// <param name="text">Message to parse</param>
+    private void ParseMessage(string text, Guid socketId)
+    {
+        try
+        {
+            Message? message = JsonSerializer.Deserialize<Message>(text)
+                               ?? throw new Exception("failed to deserialize Message");
+
+            _logger.Debug(
+                $"Received message: {(message.MessageType.Length > 32 ? string.Concat(message.MessageType.AsSpan(0, 32), "...") : message.MessageType)}"
+            );
+            _logger.Verbose(text.Length > 65536 ? string.Concat(text.AsSpan(0, 65536), "...") : text);
+
+            switch (message.MessageType)
+            {
+                case "PERFORM_ABANDON":
+                    AfterMessageReceiveEvent?.Invoke(this, new AfterMessageReceiveEventArgs(
+                        JsonSerializer.Deserialize<PerformAbandonMessage>(text)
+                        ?? throw new Exception("failed to deserialize PerformAbandonMessage"),
+                        socketId
+                    ));
+                    break;
+
+                case "PERFORM_PICK_UP":
+                    AfterMessageReceiveEvent?.Invoke(this, new AfterMessageReceiveEventArgs(
+                        JsonSerializer.Deserialize<PerformPickUpMessage>(text)
+                        ?? throw new Exception("failed to deserialize PerformPickUpMessage"),
+                        socketId
+                    ));
+                    break;
+
+                case "PERFORM_SWITCH_ARM":
+                    AfterMessageReceiveEvent?.Invoke(this, new AfterMessageReceiveEventArgs(
+                        JsonSerializer.Deserialize<PerformSwitchArmMessage>(text)
+                        ?? throw new Exception("failed to deserialize PerformSwitchArmMessage"),
+                        socketId
+                    ));
+                    break;
+
+                case "PERFORM_USE_MEDICINE":
+                    AfterMessageReceiveEvent?.Invoke(this, new AfterMessageReceiveEventArgs(
+                        JsonSerializer.Deserialize<PerformUseMedicineMessage>(text)
+                        ?? throw new Exception("failed to deserialize PerformUseMedicineMessage"),
+                        socketId
+                    ));
+                    break;
+
+                case "PERFORM_USE_GRENADE":
+                    AfterMessageReceiveEvent?.Invoke(this, new AfterMessageReceiveEventArgs(
+                        JsonSerializer.Deserialize<PerformUseGrenadeMessage>(text)
+                        ?? throw new Exception("failed to deserialize PerformUseGrenadeMessage"),
+                        socketId
+                    ));
+                    break;
+
+                case "PERFORM_MOVE":
+                    AfterMessageReceiveEvent?.Invoke(this, new AfterMessageReceiveEventArgs(
+                        JsonSerializer.Deserialize<PerformMoveMessage>(text)
+                        ?? throw new Exception("failed to deserialize PerformMoveMessage"),
+                        socketId
+                    ));
+                    break;
+
+                case "PERFORM_STOP":
+                    AfterMessageReceiveEvent?.Invoke(this, new AfterMessageReceiveEventArgs(
+                        JsonSerializer.Deserialize<PerformStopMessage>(text)
+                        ?? throw new Exception("failed to deserialize PerformStopMessage"),
+                        socketId
+                    ));
+                    break;
+
+                case "PERFORM_ATTACK":
+                    AfterMessageReceiveEvent?.Invoke(this, new AfterMessageReceiveEventArgs(
+                        JsonSerializer.Deserialize<PerformAttackMessage>(text)
+                        ?? throw new Exception("failed to deserialize PerformAttackMessage"),
+                        socketId
+                    ));
+                    break;
+
+                case "GET_PLAYER_INFO":
+                    AfterMessageReceiveEvent?.Invoke(this, new AfterMessageReceiveEventArgs(
+                        JsonSerializer.Deserialize<GetPlayerInfoMessage>(text)
+                        ?? throw new Exception("failed to deserialize GetPlayerInfoMessage"),
+                        socketId
+                    ));
+                    break;
+
+                case "GET_MAP_INFO":
+                    AfterMessageReceiveEvent?.Invoke(this, new AfterMessageReceiveEventArgs(
+                        JsonSerializer.Deserialize<GetMapMessage>(text)
+                        ?? throw new Exception("failed to deserialize GetMapInfoMessage"),
+                        socketId
+                    ));
+                    break;
+
+                case "CHOOSE_ORIGIN":
+                    AfterMessageReceiveEvent?.Invoke(this, new AfterMessageReceiveEventArgs(
+                        JsonSerializer.Deserialize<ChooseOriginMessage>(text)
+                        ?? throw new Exception("failed to deserialize ChooseOriginMessage"),
+                        socketId
+                    ));
+                    break;
+
+                default:
+                    throw new InvalidOperationException(
+                        $"Invalid message type {(message.MessageType.Length > 32 ? string.Concat(message.MessageType.AsSpan(0, 32), "...") : message.MessageType)}."
+                    );
+            }
+        }
+        catch (Exception exception)
+        {
+            _logger.Error($"Failed to parse message: {exception.Message}");
+            _logger.Debug($"{exception}");
+        }
+    }
+
+}
